Syntax and Pointers
Changed the allowable 'ident' characters so snake case 'idents' are now
allowed.
'Main' is now 'main'.

Added an extra field to the 'Machine' record to keep track of the top of
the heap.

Added 'malloc' which moves the top of the heap and returns the address
of the allocated memory region. All of these regions are tag-lengthed.

Added 'calcAddr' which takes an address and offset from the top of the
stack and calculates the resulting address.

Added some instructions that load and store values using addresses on
the top of the stack.